# packages/contracts/src/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/google/exchange": {
      "post": {
        "operationId": "AuthController_exchangeGoogleCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleExchangeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/signin": {
      "post": {
        "operationId": "AuthController_loginUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registro exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos o email ya en uso"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/complete-profile": {
      "post": {
        "operationId": "AuthController_completeProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perfil completado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteProfileResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Colibrí OS API",
    "description": "API del sistema Colibrí OS — RaaS (Reputación como Servicio)",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "GoogleExchangeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Código de un solo uso recibido en /login/google-callback?code=...",
            "example": "a1b2c3..."
          }
        },
        "required": [
          "code"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Correo electrónico del usuario registrado",
            "example": "juanperez@mail.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario",
            "example": "MiPass@123",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "fullName": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "entrepreneur",
              "mentor",
              "evaluator",
              "mecenas_semilla",
              "mecenas_fundacional",
              "mecenas_cambio",
              "admin",
              "guest"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "suspended",
              "pending_profile"
            ]
          },
          "provider": {
            "type": "string",
            "enum": [
              "local",
              "google"
            ]
          },
          "linkedinId": {
            "type": "string",
            "nullable": true
          },
          "googleId": {
            "type": "string",
            "nullable": true
          },
          "cryptoWallet": {
            "type": "string",
            "nullable": true
          },
          "credentialsWallet": {
            "type": "string",
            "nullable": true
          },
          "adnHash": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "avatar": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "non_binary",
              "other",
              "prefer_not_to_say"
            ],
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "fullName",
          "role",
          "status",
          "provider",
          "createdAt",
          "updatedAt"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Usuario logueado con éxito"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          }
        },
        "required": [
          "message",
          "user"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Correo electrónico del usuario (único)",
            "example": "juanperez@mail.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario (debe ser fuerte)",
            "example": "MiPass@123",
            "minLength": 8,
            "maxLength": 15
          },
          "confirmPassword": {
            "type": "string",
            "description": "Confirmación de la contraseña",
            "example": "MiPass@123"
          },
          "fullName": {
            "type": "string",
            "description": "Nombre completo del usuario",
            "example": "Juan Perez"
          }
        },
        "required": [
          "email",
          "password",
          "confirmPassword",
          "fullName"
        ]
      },
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Usuario registrado con éxito"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          }
        },
        "required": [
          "message",
          "user"
        ]
      },
      "CompleteProfileDto": {
        "type": "object",
        "properties": {}
      },
      "CompleteProfileResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Perfil completado con éxito"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          }
        },
        "required": [
          "message",
          "user"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token emitido en el login",
            "example": "a1b2c3..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "usuario@ejemplo.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Token recibido por email"
          },
          "newPassword": {
            "type": "string",
            "example": "NuevaPass@456"
          },
          "confirmNewPassword": {
            "type": "string",
            "example": "NuevaPass@456"
          }
        },
        "required": [
          "token",
          "newPassword",
          "confirmNewPassword"
        ]
      },
...
    }
  }
}