# src/auth/auth.controller.ts
import { BadRequestException, Body, Controller, Get, HttpCode, Post, Req, Res, UseGuards } from '@nestjs/common';
import { AuthGuard } from '@nestjs/passport';
import {
  ApiTags,
  ApiOkResponse,
  ApiCreatedResponse,
  ApiUnauthorizedResponse,
  ApiBadRequestResponse,
  ApiResponse,
} from '@nestjs/swagger';
import { AuthService } from './auth.service';
import { LoginDto } from './dto/login.dto';
import { CreateUserDto } from './dto/create.dto';
import { RefreshTokenDto } from './dto/refresh-token.dto';
import { GoogleExchangeDto } from './dto/google-exchange.dto';
import { clearAuthCookie, setAuthCookie } from './cookie.helper';
import { AuthResponseDto } from './dto/auth-response.dto';
import { RegisterResponseDto } from './dto/register-response.dto';
import { CompleteProfileResponseDto } from './dto/complete-profile-response.dto';
import { MessageResponseDto } from './dto/message-response.dto';
import type { Request, Response } from 'express';
import type { IGoogleUser } from './interfaces/googleUser.interface';
import { CompleteProfileDto } from './dto/complete-profile.dto';
import { OAuthExchangeService } from './oauth/oauth-exchange.service';
import { Throttle } from '@nestjs/throttler';
import { Ip } from '@nestjs/common';
import { ForgotPasswordDto } from './dto/forgot-password.dto';
import { ResetPasswordDto } from './dto/reset-password.dto';
import { PasswordResetService } from './password-reset/password-reset.service';

type GoogleAuthenticatedRequest = Request & { user: IGoogleUser };

@ApiTags('Auth')
@Controller('auth')
export class AuthController {
  constructor(
    private readonly authService: AuthService,
    private readonly oauthExchangeService: OAuthExchangeService,
    private readonly passwordResetService: PasswordResetService,
  ) {}

  @UseGuards(AuthGuard('google'))
  @Get('google')
  @ApiResponse({ status: 302, description: 'Redirect a Google OAuth' })
  async getGoogle() {}

  /**
   * OAUTH-001: ya no emite el JWT acá. El guard de passport valida el
   * `state` (ver OAuthStateStore) antes de llegar a este handler. Si es
   * válido, resolvemos el usuario y solo redirigimos con un código opaco
   * de un solo uso; el JWT se emite recién cuando el frontend lo canjea
   * en POST /auth/google/exchange.
   */
  @UseGuards(AuthGuard('google'))
  @Get('google/callback')
  @ApiResponse({
    status: 302,
    description: 'Redirect al frontend con cookie HttpOnly colibri_access_token',
  })
  async getGoogleCallback(@Req() req: GoogleAuthenticatedRequest, @Res() res: Response) {
  
  const result = await this.authService.googleLogin(req.user);
  const isProduction = process.env.NODE_ENV === 'production';

  if (result.requiresProfileCompletion) {
    const redirectUrl = `${process.env.FRONTEND_URL}/login/google-callback?tempToken=${result.tempToken}`;
    return res.redirect(redirectUrl);
  }
  res.cookie('colibri_access_token', result.token, {
    httpOnly: true,
    secure: isProduction,
    sameSite: isProduction ? 'none' : 'lax',
    maxAge: Number(process.env.AUTH_COOKIE_MAX_AGE_MS ?? 3_600_000),
    path: '/',
  });

  return res.redirect(`${process.env.FRONTEND_URL}/login/google-callback?role=${result.role}`);
}

  @Throttle({ default: { limit: 5, ttl: 60_000 } })
  @Post('google/exchange')
  @HttpCode(200)
  async exchangeGoogleCode(
    @Body() dto: GoogleExchangeDto,
    @Res({ passthrough: true }) res: Response,
  ) {
    const user = await this.oauthExchangeService.consume(dto.code);
    const result = this.authService.buildAuthResult(user);
    setAuthCookie(res, result.token);
    return { message: 'Sesión iniciada con éxito', user: result.user };
  }

  @Throttle({ default: { limit: 5, ttl: 60_000 } })
  @Post('signin')
  @ApiOkResponse({ type: AuthResponseDto, description: 'Login exitoso' })
  @ApiUnauthorizedResponse({ description: 'Credenciales inválidas' })
  async loginUser(@Body() logindto: LoginDto, @Res({ passthrough: true }) res: Response) {
    const result = await this.authService.loginUser(logindto);
    setAuthCookie(res, result.token);
    return { message: result.message, user: result.user };
  }

  @Throttle({ default: { limit: 5, ttl: 60_000 } })
  @Post('signup')
  @ApiCreatedResponse({ type: RegisterResponseDto, description: 'Registro exitoso' })
  @ApiBadRequestResponse({ description: 'Datos inválidos o email ya en uso' })
  async createUser(@Body() user: CreateUserDto, @Res({ passthrough: true }) res: Response) {
    const result = await this.authService.createUser(user);
    setAuthCookie(res, result.token);
    return { message: result.message, user: result.user };
  }

@Post('complete-profile')
  @ApiOkResponse({ type: CompleteProfileResponseDto, description: 'Perfil completado exitosamente' })
  async completeProfile(@Body() dto: CompleteProfileDto, @Res({ passthrough: true }) res: Response) {
    const result = await this.authService.completeProfile(dto);
    setAuthCookie(res, result.token);
    return { message: "Perfil completado con éxito", user: result.user };
  }

  @Post('refresh')
  async refresh(@Body() dto: RefreshTokenDto, @Res({ passthrough: true }) res: Response) {
    const result = await this.authService.refresh(dto.refreshToken);
    setAuthCookie(res, result.token);
    return { message: 'Tokens renovados correctamente'}
  }

  @Post('logout')
  @HttpCode(200)
  @ApiOkResponse({ type: MessageResponseDto, description: 'Sesión cerrada exitosamente' })
  logout(@Res({ passthrough: true }) res: Response) {
    clearAuthCookie(res);
    return { message: 'Sesión cerrada con éxito' };
  }

  @Throttle({ default: { limit: 5, ttl: 60_000 } }) // 5 solicitudes / min por IP
  @Post('forgot-password')
  async forgotPassword(@Body() dto: ForgotPasswordDto, @Ip() ip: string) {
    return await this.passwordResetService.requestReset(dto.email, ip);
  }

  @Throttle({ default: { limit: 5, ttl: 60_000 } })
  @Post('reset-password')
  async resetPassword(@Body() dto: ResetPasswordDto) {
    if (dto.newPassword !== dto.confirmNewPassword) {
      throw new BadRequestException('Las contraseñas nuevas deben ser iguales');
    }
    return await this.passwordResetService.resetPassword(dto.token, dto.newPassword);
  }
}